Fix functional tests
The test harness was writing the output from each test to a file via
redirection when executing the test.  However, each test was also
attempting to grep the file that was being used to redirect output,
and to redirect the output of grep to the same file.

This change fixes up trying to use the same file for input & output.