LMSBW_components: No longer an associative array
Since the concept of a 'module' corresponding to a 'component' has
been removed, there is no need to keep LMSBW_components as an
associative array; it only needs to be a list of loaded components.

This change makes the necessary changes to remove the associative
array and replace it with a simple list.