Implement source & binary API checking
This change implements API checking, both source & binary.

The checking is done during the build of the component which uses the
API.  This relationship is established through the 'prerequisite'
clause of a component definition.

Checking at the time the API client is built, while being more
resource intensive, is the only correct way of implementing the check.

Consider the following:

  o Component A exports a source & binary API.
    Component B uses both APIs.
    Component C uses both APIs.

  o There are two products:

    Product 0 contains A & B.
    Product 1 contains A & C.

  o Product 0 is being built. The API of A changes.

    If the API check is done when building A, there is no notion that
    component C is affected, because it is not included in the current
    product.

    A will be successfully built.  B will be rebuilt.

    If product 1 is now built, C may not be built because it has no
    source changes, and no imported API changes.

If build output were not shared, then API checking could be done at
the point where the exporting compnent is built, but that would mean
more compilation when switching different products.