Component 'direct-dependents' key is removed.
Calculating the direct dependents of each component at the point when
the component configuration files are loaded was taking too much time.

Since this information is only used to produce reports, the call has
been moved into the report generation module, and it is deferred to
runtime.

There is no longer a startup cost to determining the direct dependents
of components.