Renamed some component declaration functions to be 'attributes'.
This change alters the semantics of the following to be attributes of a component

   o no parallel build
   o build target
   o install target
   o api

The decision to do this is based on making it easier to search
configuration files for apsects of components which are not specific
to their declaration to LMSBW proper.  By annotating these items with
an 'attribute', it become easier to search for these items.