Rename component build targets
Sicne overriding component build targets requires makefile targets to
be placed into the component configuration file, and the component
configuration file is also loaded into the LMSBW-proper context, there
could be some confusion about rules & prerequisites in the master
Makefile.

To alleviate that, and to make it clear where the prerequiresites for
override targets originate, they are now named to be much more unique.