wrapper: {engine, component}: Fix performance regression on 'generate' sample.
Without this change, the generate sample with lmsbw takes 30 minutes,
but it should only take about 17-18 minutes.

The cause of the problem was putting the invocation of Make for the
component build & install into functions.  I do not know why this
caused such a performance degradation.

The solution is to only export the options provided to the component
Make for both build & install; these can be used when overriding the
default behavior to invoke Make the same way as lmsbw.