Implementation of 'build-output-download' components
This enables components to be marked with a 'build-output-download'
attribute.  Components thusly marked will not normally be built, but
will instead have their build output (DESTDIR) downloaded.

The download is done through a script which must be provided by the
user.  It must be provided by the user (build team) as that allows
each implementation of a build with lmsbw to be flexible -- in other
words, the author of lmsbw cannot predict how each implementation may
want to store their build output (maybe some will store raw data that
is copied with 'rsync', while others have tarballs, and still others
have a system which must use 'curl' to download).

Includes implementation, documentation, sample project and tests.