wrapper: Use rsync --checksum rather than --update
Because lmsbw does not remove build output, and because it's easy to
switch between different build directories for the same component (by
changing CFLAGS, for example), it is important that the current build
directory output is copied into the install directory.

'--update' does not do this correctly; instead use '--checksum' to
ensure that the files are properly copied.