wrapper: Added 'global-settings' & 'local-settings'
Added a 'global-settings' key to LMSBW_configuration and a
'local-settings' key to each component configuration.

Both are used to set key/value pairs that are passed on to each
component's build system.  The 'global-settings' applies to all
components, and the 'local-settings' overrides and supplements the
values of 'global-settings'.

These values are expanded into commandline 'key="value"' settings when
invoking the compnent's build process.

The 'global-settings' expansion also participates in global hashing;
changes to the global values will yield a different install directory.

The 'local-settings' participate in component hashing, and changes
will yield a different build directory.

The associative arrays can be used to control component Makefiles via
macro expansion and conditional execution.