Remove differentiation between 'image' & 'build' component directory trees
There was an original belief that separating components that are
included to help with the build, and components that are part of the
produced product would be useful; installing a whole toolchain in the
'install' directory for both types of build is an example of how it
would be convenient.

However, it turns out to not be as useful as thought.  When there are
'build' components which use the 'installed' output from 'image'
components, the system breaks down wholly.  Also, the system made it
impossible to work because a 'build' component could not be dependent
upon an 'image' component.

This change removes the distinction between the build / install
locations for 'image' and 'build' components.

It does leave intact the LMSBW_C_INSTALL_DIRECTORY and
LMSBW_C_BUILD_INSTALL_DIRECTORY macros passed to the component build
process, just in case it is desirable to separate these different
component types again.