lmsbw-mtree-check-api: Use better mtree specification keywords
For a build process, the only flags we are really concerned with are
the 'modification time' and the size of the file.  Either one of these
are enough to indicate the file is the same or the file has changed.

Using 'cksum' causes mtree to perform more work than necessary and
slowed down API checking a measurable amount.