wrapper: remove two-stage build / install
This change removes the invocation of the component's Makefile twice
(once for build, once for install) and instead invokes it once to
install.

The component Makefile must 'build' and 'install' the component.
(After all, the Makefile will not be invoked if there is nothing to
build).

The purpose of this change is a minor speed improvement: it takes
less time to start one process than it does ot start two
processes.