destdir->install: Use symlinks
Rather than use 'rsync' to copy from the component's DESTDIR to the
global install directory, this change uses the 'lndir' program.

This has been changed because copying lots of files, or a few large
files with each build is actually a time sync.  Rather than copying
the files, a symlink farm will be used, as it is much faster.