generate-component-rules.mk: Address issues with interface checking
The previous method of checking interfaces:

  short circuiting check when an out-of-date interface is found

is incorrect.  It's incorrect because all remaining interfaces are
unchecked and will only be checked (one per invocation) the next time
the build is executed.

This leads to the component build process being unnecessarily entered
multiple times.

This change restores the semantic that all interfaces must be checked.

They are &-ed off so they all run concurrently.