Improve determination of component (re-)build
This alters the way that component prerequisite interfaces are
checked.  Rather than always checking the interface (which will rarely
change in a large project), the prerequisite component will now write
data into the dependent component's build-root-directory to indicate
that the prerequisite has been updated.

This new file is checked against the timestamp of the prerequisite's
manifest in the depenedent build-root directory.  If the sentinel file
is newer, then the prerequisite is checked and the dependent's build
will be entered.

It would probably be possible to dispense entirely with the use of
mtree on prerequisite components and just re-enter the dependent's
build when a prerequisite changes (and assume that the dependent build
process is well written).

To facilitate this, each component now has a 'direct-dependent' field.
This field contains all the direct dependents of the component.