[wrapper]: Improve prerequisite checking
This change disposes with local checking of prerequisite APIs and
instead only checks if prerequisite components have been rebuilt.  If
they have been rebuilt, then the current component's build process
will be invoked.

The current component's build process should be robust enough (with
automatic dependency generation) to only rebuild the minimal set of
files when a prerequisite is changed.

The reason for this is a significant increase in NULL build speed.