[lmsbw] Add capability to adapt LMSBW build to legacy build
When using LMSBW to improve an existing build system it is seldom
possible to re-write from scratch.  This change adds the capability to
name a script that can be used to 'adapt' the output from LMSBW into
the build directory of the old build process.

You can have one adapter script declared for a single configuration.

The adapter script will be called, unconditionally, at the time the
LMSBW is complete.

The script will be provided a single command line argument that is the
pathname of a file that contains information about each component that
has been built by LMSBW.

   component-file := (component-info '\n')...
   component-info := component-name ':' component-build-root-directory

The 'component-name' is the name of the component as declared to
LMSBW.

The 'component-build-root-directory' is the full path to the root of
the build directory for the component as configured.  The build-root
directory contains a 'build' and 'destdir' directory.  The 'destdir'
directory contains the exported build output that was installed by the
component's build process.

Because the adapter script is unconditionally executed, it is
advisable to write a sentinel file into the component's
build-root-directory to make it so the component build output is only
adapted to the old build process when it actually changes.