[attribute functions] Added 'exports api' attribute
This attribute indicates that the component has an API that needs to
be installed into the global install directory before any dependent
component, or itself, can be built.

This is the first in a series of changes that will obviate the need to
make a logical single copmonent into two -- one which installs the
headers and one which builds the component.