[wrapper] Move sentinel files in preparation for deleted file fix
Sentinel files stored directly in the build directory will be deleted
when the build & destdir directores need to be deleted before syncing
a component's sources to the build directory.

The component's build & destdir directories need to be deleted before
syncing the component sources when files have been deleted from the
component source directory.

Failure to delete the files can yield bad builds.

This change moves the sentinel files so they will not be affected by a
follow-on change to deal with files deleted from the component source
directory.