[scripts] Fix Manifest Checking
A recent discovery shows that mtree does not return a non-zero error
code (indicating a mismatch between the current directory and the
stored manifest when files are deleted from the directory.

Instead, mtree prints, to stdout, that files have been deleted and
returns 0 if no other mismatches occur.

This change fixes this deficiency of mtree by checking for deleted
files and properly reporting that the component has changed.

This is a stand-alone precursor for detecting when files have been
deleted from the component's source directory and automatically
deleting the build & destdir directories before syncing the source.