[wrapper] Correctly handle files deleted from source directories
If files are deleted from a component's source directory, and the
component has already been built, the deleted files will persist in
the build directory; this will lead to bad builds.

This change adds and additional step to syncing the source directory
to the build directory: a pre-sync check.

The pre-sync action checks if any files have been removed from the
source directory, and if they have, the build & destdir directories
are removed and recreated.

This ensures that no stale files will be left over in the build
directory before the component is produced.

Files installed in the global install directory will also be removed.