[lmsbw] Custom Toolchains
Adds support for custom toolchains.

Custom toolchains allow the implementor of a build system to use
toolchains that have not been created with ctng.  The developer
provides a Make include file that is when ctng toolchains are not
used, or not found.  The developre will examind the LMSBW_C_TOOLCHAIN
macro to determine the name of the toolchain to be used, and then set
the standard Make macros accordingly.

If the custom toolchain include file does not recognize the toolchain,
it must produce an error with $(error).